init: move all manifest updating code to `manifest_update_scheduler.ts`
